src/auth/backend/requirements.txt
fastapi 
uvicorn 
python-jose[cryptography] 
passlib[bcrypt] 
python-dotenv 
pydantic 
# dotenv 
# jose
sqlalchemy
psycopg2-binary
pytest 
httpx